update 2_get_clustering.R
changing so that the clustering is performed on a log10 transformed version of all datasets (as done in Davidson 2024 - https://github.com/greenelab/hgsc_characterization/blob/master/figure_notebooks/compare_centroids.Rmd). Prior code was running clustering on the _filtered_transformed versions of all datasets - this is in log10 for rnaseq datasets and in pseudo-counts for microarray datasets.